only removing the tmp datajockey beats file if it exists now we can recursively add files to import, the import_audio script has commandline argument options
git-svn-id: svn+ssh://yew/home/alex/code-repo/projects/datajockey/code/qtgui@845 be07c712-5235-0410-95d4-c99efef06bb2

+++ ruby/datajockey/annotation/beatlocations.rb
@@ -30,8 +30,12 @@ module Datajockey
         unless system("java", "-jar", @@beatroot_jarfile_loc, "-o", a.path, tempwav)
           raise "Cannot execute beatroot, is the jar file in a known location?"
         end
+      rescue => e
+        raise e
       ensure
-        FileUtils.rm(tempwav)
+        if File.exists?(tempwav)
+          FileUtils.rm(tempwav)
+        end
       end
       return a.readlines.collect{|l| l.to_f}
     end
